v0.8.3 release

+++ vllm/v1/executor/multiproc_executor.py
@@ -52,7 +52,11 @@ class MultiprocExecutor(Executor):
                 "shutting down. See stack trace above for root cause issue.")
             # Propagate error up to parent process.
             parent_process = psutil.Process().parent()
-            parent_process.send_signal(signal.SIGUSR1)
+            if parent_process:
+                if platform.system() == "Windows":
+                    parent_process.send_signal(signal.SIGTERM)
+                else:
+                    parent_process.send_signal(signal.SIGUSR1)
             self.shutdown()
 
         if platform.system() != "Windows":
